Height and Width syntax
Height and width of objects are defined with vh and vw for optimimum responsiveness of screen size.
Text of objects and all new objects will need to be made appropriate size. Currently the screen size is 950px base...but this needs to be bigger and will fix that currently. Main focus of this was to get layout set responsive.